HomePage컴포넌트에서 배너 컴포넌트 추출
- component/page/Banner.js 생성
