+++ CODEOWNERS
@@ -12,3 +12,6 @@
 
 # HomePage.js 파일 소유
 HomePage.js @rkdgusrn1212
+
+# page/Banner.js 파일 소유
+page/Banner.js @rkdgusrn1212
